[FE] feat: 토큰이 있을 시 교환&나눔 게시판 이용 가능
