--- src/assets/fonts/fonts.css
@@ -1,10 +1,0 @@
-@font-face {
-    font-family: '210 옴니고딕OTF 045';
-    src: url('./210 옴니고딕OTF 045.otf') format('opentype');
-}
-@font-face {
-    font-family: 'Gugi';
-    src: url('Gugi-Regular.ttf') format('truetype');
-    font-weight: normal;
-    font-style: normal;
-}
